Move t.photo-related classes to takephoto package
