+++ app/src/main/java/com/shapeapp/shape/ui/takephoto/TakePhotoFragmentViewModel.kt
@@ -1,4 +1,4 @@
-package com.shapeapp.shape.ui
+package com.shapeapp.shape.ui.takephoto
 
 import android.net.Uri
 import androidx.lifecycle.LiveData
