+++ app/src/main/res/layout/fragment_take_photo.xml
@@ -5,7 +5,7 @@
     android:id="@+id/root_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.TakePhotoFragment">
+    tools:context=".ui.takephoto.TakePhotoFragment">
 
     <ImageView
         android:id="@+id/taken_photo_imageview"
